refactor(examples): dual-target (GJS + browser) for WebRTC examples
Restructure webrtc-loopback, webrtc-video, and webtorrent-seed to the
standard dual-target pattern: shared platform-agnostic logic extracted
to a separate module, with gjs/ and browser/ entry points.

Pattern (matching canvas2d-confetti, three-geometry-cube, etc.):
  src/<shared>.ts     — platform-agnostic logic (LogFn callback)
  src/gjs/gjs.ts      — GLib.MainLoop + print() logging
  src/browser/browser.ts — DOM <pre> logging / <video> element
  src/browser/index.html — minimal HTML shell

webrtc-loopback: loopback-demo.ts exports runLoopback(log)
  GJS: tested ✓ — full data channel echo (string + binary)
  Browser: build ✓ — uses native RTCPeerConnection

webrtc-video: video-demo.ts exports startVideo(video, log)
  GJS: VideoBridge + Adwaita window
  Browser: native <video> + getUserMedia

webtorrent-seed: torrent-demo.ts exports runTorrentDemo(log)
  GJS: --globals auto,RTCPeerConnection,... for WebTorrent
  Browser: --app gjs --globals none (Node polyfills needed)

Build scripts: build:gjs, build:browser, build:assets, start:browser
