+++ examples/dom/webrtc-loopback/src/browser/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>WebRTC Data-Channel Loopback</title>
+    <style>
+        body { margin: 0; padding: 1rem; font-family: monospace; background: #1e1e2e; color: #cdd6f4; }
+        h1 { font-size: 1.2rem; color: #89b4fa; }
+        #log { white-space: pre-wrap; font-size: 0.85rem; line-height: 1.5; max-height: 80vh; overflow-y: auto; }
+    </style>
+</head>
+<body>
+    <h1>WebRTC Data-Channel Loopback</h1>
+    <pre id="log"></pre>
+    <script defer src="browser.js"></script>
+</body>
+</html>
