+++ examples/dom/webrtc-video/src/browser/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>WebRTC Video — Webcam Preview</title>
+    <style>
+        body { margin: 0; background: #1e1e2e; display: flex; flex-direction: column; align-items: center; justify-content: center; min-height: 100vh; }
+        video { max-width: 100%; max-height: 80vh; border-radius: 12px; background: #000; }
+        #status { color: #cdd6f4; font-family: monospace; margin-top: 1rem; font-size: 0.9rem; }
+    </style>
+</head>
+<body>
+    <video autoplay playsinline muted></video>
+    <div id="status">Initializing...</div>
+    <script type="module" src="browser.js"></script>
+</body>
+</html>
