+++ examples/dom/webtorrent-seed/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@gjsify/example-dom-webtorrent-seed",
-  "description": "WebTorrent local seed + download demo — tests @gjsify/webrtc data channels via P2P torrent transfer in a single GJS process",
-  "main": "dist/index.gjs.js",
+  "description": "WebTorrent local seed + download demo — tests WebRTC data channels via P2P torrent transfer for GJS and browser",
+  "main": "dist/gjs.js",
   "type": "module",
   "engines": {
     "node": ">=22.12.0"
@@ -10,10 +10,12 @@
     "clear": "rm -rf dist tsconfig.tsbuildinfo",
     "check": "tsc --noEmit",
     "start": "yarn start:gjs",
-    "start:gjs": "gjsify run dist/index.gjs.js",
-    "build": "yarn build:gjs",
-    "build:gjs": "gjsify build src/index.ts --app gjs --outfile dist/index.gjs.js --globals auto,RTCPeerConnection,RTCSessionDescription,RTCIceCandidate,WebSocket,fetch",
-    "postbuild:gjs": "node -e \"const fs=require('fs');let c=fs.readFileSync('dist/index.gjs.js','utf8');c=c.replaceAll('__toESM(require_lib3(), 1)','__toESM(require_lib3(), 0)');fs.writeFileSync('dist/index.gjs.js',c)\""
+    "start:gjs": "gjsify run dist/gjs.js",
+    "start:browser": "http-server dist -o -c-1",
+    "build": "yarn build:gjs && yarn build:browser && yarn build:assets",
+    "build:gjs": "gjsify build src/gjs/gjs.ts --app gjs --outfile dist/gjs.js --globals auto,RTCPeerConnection,RTCSessionDescription,RTCIceCandidate,WebSocket,fetch",
+    "build:browser": "gjsify build src/browser/browser.ts --app gjs --outfile dist/browser.js --globals none",
+    "build:assets": "cp -f src/browser/index.html dist/index.html"
   },
   "devDependencies": {
     "@girs/gjs": "^4.0.0-rc.3",
@@ -24,6 +26,7 @@
     "@gjsify/webrtc": "workspace:^",
     "@gjsify/websocket": "workspace:^",
     "@types/node": "^25.6.0",
+    "http-server": "^14.1.1",
     "typescript": "^6.0.2",
     "webtorrent": "^2.5.18"
   },
